apps/server/scripts/requirements.txt
google-cloud-bigquery==2.30.1
psycopg2-binary==2.9.3
pyarrow==14.0.1
numpy==1.24.3
uuid
pandas
open-rarity
sqlalchemy
psycopg2-binary
openpyxl
python-dotenv>=0.19.0
tqdm 